+++ receiver/nmea/receiver.cpp
@@ -39,7 +39,7 @@ void NmeaReceiver::process() {
             }
 
             // Ignore the return value, this should never fail if the parser buffer is >= 1024 bytes
-            mParser->append(buffer, length);
+            mParser->append(buffer, static_cast<uint16_t>(length));
         }
     }
 }
